+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKH5NDG4P1VFMF23SBQE69KN"
+        "01KKJ1Q0BKV0T4K9B9044ESXVH"
     )
     r"""The unique identifier of the data source"""