+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -64,7 +64,7 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-03-12T14:01:50.585Z")
+    created: Optional[datetime] = parse_datetime("2026-03-12T22:12:00.933Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")