+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKH5NDGDQWBP1BKXGZ0FH7GN"
+        "01KKJ1Q0BNYBZWP1XXJDFY6PRT"
     )
     r"""The unique identifier of the data source"""